rename init to registerTranslations
