+++ src/app/static/src/app/forgotPassword.ts
@@ -5,7 +5,7 @@ import Vuex from "vuex";
 import {initialPasswordState, PasswordState} from "./store/password/password";
 import {actions} from './store/password/actions';
 import {mutations} from './store/password/mutations';
-import init from "./store/translations/init";
+import registerTranslations from "./store/translations/registerTranslations";
 
 Vue.use(Vuex);
 
@@ -16,7 +16,7 @@ const passwordStoreOptions: StoreOptions<PasswordState> = {
 };
 
 const store = new Vuex.Store<PasswordState>(passwordStoreOptions);
-init(store);
+registerTranslations(store);
 
 export const forgotPasswordnApp = new Vue({
     el: "#app",
