+++ src/app/static/src/app/main.ts
@@ -1,9 +1,9 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
 import {RootState, storeOptions} from "./root";
-import init from "./store/translations/init";
+import registerTranslations from "./store/translations/registerTranslations";
 
 Vue.use(Vuex);
 
 export const store = new Vuex.Store<RootState>(storeOptions);
-init(store);
+registerTranslations(store);
