+++ src/app/static/src/app/store/translations/registerTranslations.ts
